Refactor: Modal Component Common으로 이동
