+++ client/src/components/Common/Modal/index.style.tsx
