+++ client/src/components/SignInModal/index.tsx
@@ -1,4 +1,4 @@
-import Modal from '@components/Modal/index';
+import Modal from '@components/Common/Modal/index';
 import { ReactComponent as GoogleIcon } from '@assets/icons/google.svg';
 import { ReactComponent as NaverIcon } from '@assets/icons/naver.svg';
 import { ReactComponent as FacebookIcon } from '@assets/icons/facebook.svg';
